run test with custom workflow engine (#3278)
* run test with custom workflow engine


* test list and task list finished


* debug


* debug


* debug


* debug


* debug


* fix missing api changes


* fix compilation


* debug


* debug


* add missing apis


* add field to test detail


* fix testing logs error


* debug logs


* debug


* debug logs


* debug


* debug


* debug


* debug


* debug


* debug


* debug


* scanning task fix


* fix logs


* fix sse logs


* debug


* change response in list custom workflow api


---------

+++ pkg/microservice/aslan/core/workflow/testing/service/types.go
@@ -138,7 +138,7 @@ type ScanningStatistic struct {
 type ListScanningTaskResp struct {
 	ScanInfo   *ScanningInfo       `json:"scan_info"`
 	ScanTasks  []*ScanningTaskResp `json:"scan_tasks"`
-	TotalTasks int                 `json:"total_tasks"`
+	TotalTasks int64               `json:"total_tasks"`
 }
 
 type ScanningInfo struct {
